+++ community/kernel/src/main/java/org/neo4j/kernel/api/ReadOperations.java
@@ -185,7 +185,7 @@ public interface ReadOperations
     long nodeGetFromUniqueIndexSeek( IndexDescriptor index, Object value ) throws IndexNotFoundKernelException,
             IndexBrokenKernelException;
 
-    int nodesCountIndexed( IndexDescriptor index, long nodeId, Object value )
+    long nodesCountIndexed( IndexDescriptor index, long nodeId, Object value )
             throws IndexNotFoundKernelException, IndexBrokenKernelException;
 
     boolean nodeExists( long nodeId );
